Fix progress bar exceeding 100%
When running the fit in `challenge2.ipynb`, the progress bar is configured for a maximum of 21 items, but the actual fitting loop processes 42 items (each descriptor is fit twice, once for mean and once for standard deviation).  This causes the progress bar to appear to exceed 100% as it progresses past the 21st item.  Changing the scale of the progress bar to `n_descriptors * 2` should bring it into agreement with reality.